Change RunnerId/AgentId from int32 to uint64 (#2661)
* RunnerId int -> ulong

+++ src/Runner.Listener/SelfUpdater.cs
@@ -30,7 +30,7 @@ namespace GitHub.Runner.Listener
         private ITerminal _terminal;
         private IRunnerServer _runnerServer;
         private int _poolId;
-        private int _agentId;
+        private ulong _agentId;
 
         public bool Busy { get; private set; }
